+++ pythonbird/__init__.py
@@ -1,10 +1,16 @@
+from .client import Thunderbird
 from .contacts import ThunderbirdContacts, ThunderbirdContactsError
 from .core import ThunderbirdLinux
 from .mail import ThunderbirdMail
+from .models import Attachment, Contact, Message
 
 __all__ = [
+    "Thunderbird",
     "ThunderbirdLinux",
     "ThunderbirdMail",
     "ThunderbirdContacts",
     "ThunderbirdContactsError",
+    "Message",
+    "Attachment",
+    "Contact",
 ]